exec: add sorter
```
BenchmarkSort/rows=2048/cols=1-8                    5000            287366 ns/op          57.01 MB/s
BenchmarkSort/rows=2048/cols=2-8                    5000            292186 ns/op         112.15 MB/s
BenchmarkSort/rows=2048/cols=4-8                    5000            318730 ns/op         205.62 MB/s
BenchmarkSort/rows=16384/cols=1-8                   1000           2270545 ns/op          57.73 MB/s
BenchmarkSort/rows=16384/cols=2-8                    500           2425301 ns/op         108.09 MB/s
BenchmarkSort/rows=16384/cols=4-8                    500           2692579 ns/op         194.72 MB/s
BenchmarkSort/rows=262144/cols=1-8                    30          34905432 ns/op          60.08 MB/s
BenchmarkSort/rows=262144/cols=2-8                    30          38191955 ns/op         109.82 MB/s
BenchmarkSort/rows=262144/cols=4-8                    30          47341978 ns/op         177.19 MB/s
```

It appears we're fully bound on Go's sort implementation. I suspect that
removing the n log n interface lookups would speed this up a bunch. It's
annoying that Go doesn't have templated sort.


Release note: None

+++ pkg/sql/exec/aggregator.go
@@ -222,7 +222,10 @@ func (a *orderedAggregator) Next() ColBatch {
 		for i := 0; i < len(a.aggTyps); i++ {
 			// According to the aggregate function interface contract, the value at
 			// the current index must also be copied.
-			a.scratch.ColVec(i).Copy(a.scratch.ColVec(i), a.scratch.outputSize, a.scratch.resumeIdx+1, a.aggTyps[i][0])
+			a.scratch.ColVec(i).Copy(a.scratch.ColVec(i),
+				uint64(a.scratch.outputSize),
+				uint64(a.scratch.resumeIdx)+1,
+				a.aggTyps[i][0])
 			a.scratch.resumeIdx = a.scratch.resumeIdx - a.scratch.outputSize
 			if a.scratch.resumeIdx >= a.scratch.outputSize {
 				// We still have overflow output values.
